refactor: make the hook more generic
