+++ packages/twenty-ui/src/display/icon/components/TablerIcons.ts
@@ -125,6 +125,7 @@ export {
   IconDotsVertical,
   IconDownload,
   IconEditCircle,
+  IconExclamationCircle,
   IconExternalLink,
   IconEye,
   IconEyeOff,
