+++ packages/twenty-ui/src/display/index.ts
@@ -186,6 +186,7 @@ export {
   IconDotsVertical,
   IconDownload,
   IconEditCircle,
+  IconExclamationCircle,
   IconExternalLink,
   IconEye,
   IconEyeOff,
